Test and script to analyze the kube-proxy iptables performance improvements

## How to run it

```
cd (kubernetes_repo)/pkg/proxy/iptables
# modify the pkg/proxy/iptables/proxier_performance_test.go file accordingly
# to define the workload #services, endpoints per service and rate to create and delete
./run_performance_test.sh
```

The script creates a network namespace, compiles the test and run the test binary inside the namespace, this way the host is not polluted with iptables rules.

### Example

```
$ ./run_performance_test.sh 
I0914 17:39:48.882291 1960435 proxier.go:235] "Setting route_localnet=1, use nodePortAddresses to filter loopback addresses for NodePorts to skip it https://issues.k8s.io/90259"
I0914 17:39:48.882404 1960435 utils.go:431] "Changed sysctl" name="net/ipv4/conf/all/route_localnet" before=0 after=1
I0914 17:39:48.882421 1960435 proxier.go:251] "Using iptables mark for masquerade" ipFamily=IPv4 mark="0x00004000"
I0914 17:39:48.882443 1960435 proxier.go:295] "Iptables sync params" ipFamily=IPv4 minSyncPeriod="1s" syncPeriod="30s" burstSyncs=2
I0914 17:39:48.882468 1960435 proxier.go:305] "Iptables supports --random-fully" ipFamily=IPv4
I0914 17:39:48.882475 1960435 proxier.go:235] "Setting route_localnet=1, use nodePortAddresses to filter loopback addresses for NodePorts to skip it https://issues.k8s.io/90259"
I0914 17:39:48.882501 1960435 proxier.go:251] "Using iptables mark for masquerade" ipFamily=IPv6 mark="0x00004000"
I0914 17:39:48.882517 1960435 proxier.go:295] "Iptables sync params" ipFamily=IPv6 minSyncPeriod="1s" syncPeriod="30s" burstSyncs=2
I0914 17:39:48.882529 1960435 proxier.go:305] "Iptables supports --random-fully" ipFamily=IPv6
I0914 17:39:48.882542 1960435 proxier.go:796] "Not syncing iptables until Services and Endpoints have been received from master"
I0914 17:39:48.882549 1960435 proxier.go:796] "Not syncing iptables until Services and Endpoints have been received from master"
I0914 17:39:48.882558 1960435 proxier.go:839] "Syncing iptables rules"
I0914 17:39:48.941995 1960435 proxier.go:1489] "Reloading service iptables data" numServices=0 numEndpoints=0 numFilterChains=5 numFilterRules=3 numNATChains=4 numNATRules=5
I0914 17:39:48.944281 1960435 proxier.go:804] "SyncProxyRules complete" elapsed="61.722221ms"
I0914 17:39:48.944296 1960435 proxier.go:839] "Syncing iptables rules"
I0914 17:39:48.999910 1960435 proxier.go:1489] "Reloading service iptables data" numServices=0 numEndpoints=0 numFilterChains=5 numFilterRules=3 numNATChains=4 numNATRules=5
I0914 17:39:49.002166 1960435 proxier.go:804] "SyncProxyRules complete" elapsed="57.861513ms"
2022-09-14 17:39:49.002265829 +0200 CEST m=+0.137106634: rules: 5.00 svc: 0.00 eps: 0.00 restoreFailures: 0.00 
Adding services
I0914 17:39:49.017937 1960435 proxier.go:839] "Syncing iptables rules"
Adding endpoints
I0914 17:39:49.057451 1960435 proxier.go:1489] "Reloading service iptables data" numServices=123 numEndpoints=0 numFilterChains=5 numFilterRules=126 numNATChains=4 numNATRules=5
I0914 17:39:49.083567 1960435 proxier.go:804] "SyncProxyRules complete" elapsed="66.57412ms"
I0914 17:39:49.279406 1960435 proxier.go:839] "Syncing iptables rules"
I0914 17:39:49.506904 1960435 proxier.go:1489] "Reloading service iptables data" numServices=1000 numEndpoints=100000 numFilterChains=5 numFilterRules=3 numNATChains=101004 numNATRules=302005
2022-09-14 17:39:50.003077985 +0200 CEST m=+1.137918795: rules: 302005.00 svc: 2000.00 eps: 1000.00 restoreFailures: 0.00 
2022-09-14 17:39:51.003272045 +0200 CEST m=+2.138112853: rules: 302005.00 svc: 2000.00 eps: 1000.00 restoreFailures: 0.00 
2022-09-14 17:39:52.003639686 +0200 CEST m=+3.138480498: rules: 302005.00 svc: 2000.00 eps: 1000.00 restoreFailures: 0.00 
2022-09-14 17:39:53.004042089 +0200 CEST m=+4.138882898: rules: 302005.00 svc: 2000.00 eps: 1000.00 restoreFailures: 0.00 
2022-09-14 17:39:54.004465331 +0200 CEST m=+5.139306135: rules: 302005.00 svc: 2000.00 eps: 1000.00 restoreFailures: 0.00 
Removing services
Removing endpoints
2022-09-14 17:39:55.005233358 +0200 CEST m=+6.140074169: rules: 302005.00 svc: 4000.00 eps: 2000.00 restoreFailures: 0.00 
I0914 17:39:55.612430 1960435 trace.go:205] Trace[1298498081]: "iptables restore" (14-Sep-2022 17:39:49.506) (total time: 6105ms):
Trace[1298498081]: [6.105467313s] [6.105467313s] END
I0914 17:39:55.612463 1960435 proxier.go:804] "SyncProxyRules complete" elapsed="6.528875487s"
I0914 17:39:55.798498 1960435 proxier.go:839] "Syncing iptables rules"
2022-09-14 17:39:56.006223634 +0200 CEST m=+7.141064440: rules: 302005.00 svc: 4000.00 eps: 2000.00 restoreFailures: 0.00 
2022-09-14 17:39:57.006313042 +0200 CEST m=+8.141153847: rules: 302005.00 svc: 4000.00 eps: 2000.00 restoreFailures: 0.00 
2022-09-14 17:39:58.007245253 +0200 CEST m=+9.142086063: rules: 302005.00 svc: 4000.00 eps: 2000.00 restoreFailures: 0.00 
2022-09-14 17:39:59.008160532 +0200 CEST m=+10.143001342: rules: 302005.00 svc: 4000.00 eps: 2000.00 restoreFailures: 0.00 
Adding services
Adding endpoints
2022-09-14 17:40:00.008771369 +0200 CEST m=+11.143612179: rules: 302005.00 svc: 6000.00 eps: 3000.00 restoreFailures: 0.00 
2022-09-14 17:40:01.009510449 +0200 CEST m=+12.144351259: rules: 302005.00 svc: 6000.00 eps: 3000.00 restoreFailures: 0.00 
2022-09-14 17:40:02.009756262 +0200 CEST m=+13.144597066: rules: 302005.00 svc: 6000.00 eps: 3000.00 restoreFailures: 0.00 
2022-09-14 17:40:03.010112163 +0200 CEST m=+14.144952981: rules: 302005.00 svc: 6000.00 eps: 3000.00 restoreFailures: 0.00 
I0914 17:40:03.474844 1960435 trace.go:205] Trace[2019727887]: "iptables save" (14-Sep-2022 17:39:55.880) (total time: 7593ms):
Trace[2019727887]: [7.593844981s] [7.593844981s] END
I0914 17:40:03.542338 1960435 proxier.go:1489] "Reloading service iptables data" numServices=0 numEndpoints=0 numFilterChains=5 numFilterRules=3 numNATChains=101004 numNATRules=101005
2022-09-14 17:40:04.010784029 +0200 CEST m=+15.145624838: rules: 101005.00 svc: 6000.00 eps: 3000.00 restoreFailures: 0.00 
Removing services
Removing endpoints
2022-09-14 17:40:05.011234464 +0200 CEST m=+16.146075268: rules: 101005.00 svc: 8000.00 eps: 4000.00 restoreFailures: 0.00 
```